+++ gradle.properties
@@ -10,10 +10,10 @@ platformType = IC
 platformVersion = 2020.3.2
 platformDownloadSources = true
 
-mapperVersion = 0.1.1
-markerVersion = 0.1.1
-mentorVersion = 0.1.1
-portalVersion = 0.1.1
-protocolVersion = 0.1.1
+mapperVersion = 0.1.2
+markerVersion = 0.1.2
+mentorVersion = 0.1.2
+portalVersion = 0.1.2
+protocolVersion = 0.1.2
 
 monitorSkyWalkingVersion = 0.1.2
