+++ plugin/jetbrains/CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Upgraded
 - Apache SkyWalking (8.2.0 -> 8.3.0)
 - Vert.x (3.9.4 -> 4.0.0)
+- Kotlin (1.4.10 -> 1.4.21)
 
 ## [0.1.1] - 2020-12-01
 ### Fixed
