+++ plugin/jetbrains/src/main/resources/messages/PluginBundle_zh.properties
@@ -1,0 +1,28 @@
+host=主办最近3小时
+port=端口
+root_source_package=根源包
+version=版
+apache_skywalking_settings=Apache SkyWalking 设置
+plugin_settings=插件设置
+
+last_occurred=上次发生
+threw=扔了
+ago=前
+
+#Portal
+Entry\ method=进入方法
+Auto-subscribe=自动订阅
+Name=名称
+Type=类型
+Throughput=通量
+Response=响应
+SLA=生菜
+
+Latest\ Traces=最新痕迹
+Slowest\ Traces=最慢的痕迹
+Failed\ Traces=跟踪失败
+
+LAST\ 5\ MINUTES=最近5分钟
+LAST\ 30\ MINUTES=最近30分钟
+LAST\ HOUR=上一个小时
+LAST\ 3\ HOURS=最近3小时
